Add a new solver that is immensely faster
Rather than starting with a clean slate per row, the new solver starts
with a set of rules and builds UP toward more and more complicated
board positions, never visiting one it's already seen.  This dynamic
programming allows for re-use of all the previously viewed solutions,
and avoiding recomputation of the same solutions.

It also means we need a new solver for each set of rules, but that's
entirely ok.  For a complicated set of rules, the generator using the
solver to verify took 40 seconds.  With the new solver, it takes 0.5
seconds.

There is no support for wildcards.  The solver is not integrated yet
either, except into the generator.  (Which is also not integrated.)