Fix wildcard hints
Fairly substantial data plubmbing update, to pass through hashes
instead of arrays, where a "move" is now a hash with more info:

index     offset into gamestate string)
pat       resolved pattern the rule will transorm ("from")
repl      resolved pattern of the replacement the rule will change "to"
capture   the capture groups from wildcards (as a string)
rawpat    unresolved pattern from
rawrepl   unresolved pattern to

resolved refers to wildcards.
If the rule is "a.bc."  ; raw pat
And repl is "a2bc1"     ; raw repl

If matches against this string "aabcc"
pat would resolve to "aabcc"
and repl would resolve to "acbca"
and capture would be "ac"

Captures are ac because that's what matched the "."s in the raw pat,
when applied against the string.

When giving hints, the problem was, the hint system only saw the
resolved strings, and it was difficult to reliably match them back to
the rules, which only had the raw strings.  So with more info, it's
easier to match (raw against raw) to guarantee an unambiguous match.

Ambiguity:
2 rules:
  "..." -> ""
  "abc" -> ""

if resolved string is "abc" it matches both rules.  But now we have
the unresolved string and "..." is now known in the "move", so we
can match rules with that.

tests still need updating