fix reset bug, and pause music bug
If a playarea-grid is selected while hitting 'r' to restart the level,
it did not clear out the selection, and allowed one to apply the
previously selected move on the current board--even if it was invalid
after resetting the level, and would crash.

Bug 2: pausing the music, if left long enough, would still advance and
start the next track.  Now update() is ignored if it's paused.