Fix hint bug with wildcards
When the hint system knew the move to make, it was resolved.  That is,
a rule with wildcards and placeholders:

   "a.a" -> ["1"]

   'a-<any>-a can be replaced with <any> - whatever it matched.

Problem was the hint saw the move for a real board position:

   Position: "aba" applying rule: "a.a" -> ["1"], result: "b"

   Solver saw the move was "aba" -> "b", but when finding the replacement
   string, it didn't find "b" in the rules.  It needed to search for "1"
   instad, before the wildcard was replaced.

Solution: the gamestate generating all possible moves will include not just
the "resolved" replacement string, but the original unresolved replacement
string.  So the hint system doesn't need to do anything but read it from a
different field.

Fixed gamestate test suite to account for raw replacements in moves.