RotatingColorWigit cell now in replacement rows, and it updates
Added better z-handling to all the custom wigits, but especially
the rotating color cell, so that the color circles overlay in a
way making it clearer which is current, and which is next.

Needed to collect the rule-cells that required being updated after
each play in the game, so that the colors could visually change in
the rule too.

Factored out the CellFactory into its own file, shared by both the
level (for playarea rows) and the rules, since they each display their
cells slightly differently.

TODO: make the cell factory work for the wildcards too, and other custom
cells.  Currently there is special handling for wildcards, placeholders,
empty cell display, etc.  Those seem they could be from the factory too
and have no special casing in the code that uses them.