Refactor UI states to be abstracted
Now input_sate does basic bookkeeping and defers to the "real" state
the mouse and key events.  This will allow for other screens, menus, etc,
so they can get the input and the game is "paused".

Also, renamed the game_state.rb to game_data.rb, and all associated
variables likewise.  This is the underlying data model of the board.
The old name somewhat was obscured by the different meaning of "state"
for the UI the datamodel.