Initial compress_vertices impl, test passes
This is a nice milestone, as the vertices are now starting to
compress.  This means that the initial one-vertex-per-block can
combine, to make the size smaller, and possibly the search space.

In order for two vertices to merge, the following must be true:

1) they both have the same color

2) the "parent" node has room for some (preferably all) of the blocks
in the "child" vertex.

3) the parent is the only edge into the child

If we merge two vertices, the following happens:

* blocks from the child are put into the parent, as many as possible,
  keeping the same order

* the child loses all the blocks moved into the parent.  If the child
  becomes empty it is removed.

* to remove the child:

  * its name and Vetex is removed from Vertices
  * Vertices fills the hole with the last name and vertex
  * adjacency matrix:
    + removes edge from parent to removed child
    + gives the out-edges from removed child to its parent
    + changes edges from all parent nodes of the moved vertex to its
      new position.
    + Note: adjacency matrix size does not change, but there is an
      unused item left in it at the end.
  * all of the above is coordinated by the phase1 GraphCreator

When phase1 is complete, a new graph will be created for phase 2,
which renames nodes, removes dead cells, and is in final form to
iterate over for the isomorphism matching to begin.