Adding new color for ROTATING_COLOR
recognized a problem with the idea behind the custom color "per
rotating color" type defined.  (Note, two uses of word "color" here,
"rotating color" it refers to color blocks in the special in-game
block type, while the Colors enum refers to graph vertex color, as
used as a generic attribute on vertices.

The problem is, if each unique combination of block colors dynamically
creates a new vertex color, then it will make it very hard to
determine if two graphs that are isomorphisms but use different colors
for corresponding blocks will be detectible, because we will only
iterate/search colors for like-colored vertices, and if their vertex
color changes depending on block colors contained within it, we wont
compare properly.

Solution: All cycling color vertices will have the SAME vertex color,
now given a well defined name ROTATING_COLORS.  All blocks of this
color can have its internal simple blocks compared.  It is, however,
important to NOT merge data from these vertices, since they are
different, and the data blocks within it are the colors it cycles
through, not logical vertices themselves.  It thus always represents a
SINGLE block in the game even if there are multiple colors in it--
those are just the colors it will cycle through.

This commit simply adds the new color and updates the tests to make it
exist, and adds a new function to answer the question: "Are vertices
of this color eligible for merging?"