vertices are vector, chk mergeable by color
to avoid running out of space for a large graph, and for moving
vectors quickly (GraphCreator should be able to creat a graph by
moving its accumulator internal types into final graph), now push_back
into a vector instead of indexing into a fixed-size 32 element array
of vertex.

Add logic to ensure that vectors can only merge if their colors are
eligible and that is: two vertices are the same FinalColor (color and
side) *and* those colors are allowed to merge.  Currently only these
types are mergeable:

  1) SOLID_RECTANGLE
  2) BACKREF
  3) WILDCARD