Gut Graph class; GraphCreator creates one
The creator is incrementally building up state, and the Graph is the
final graph and optimized nodes.  (Ideally, though in this versino
Graph merely holds some data.)

But this is the first time that the GraphCreator actually creates a
graph, and that allows the next step to begin: iterating two graphs to
search isomorphisms!