AdjacencyMatrix better printing and resize_down
Now we can print the matrix with labels, which allows the vertex names
to be seen.  Or we can just print the idx.  Printing moved to separate
file (AdjacencyMatrixPrinter), and uses the (newly imported) `fmt`
library.

This was inspired by the other new AdjacencyMatrix feature,
resize_down, which can eliminate extra empty rows and columns at the
rightmost/bottommost area of the graph.  It asserts they are empty.