Cleanup; resize matrix after compression
Remove spurious output to cout

Add accessor to GraphCreator to get a matrix::WithNames to
pretty-print matrix as introducded last commit

Make names be range-checked when printing matrix

Compressing the vertices (which combines nodes and may reduce total
vertex count) removed names from the names list but the adjacency
matrix did not shrink.  Until recently there was no clean way to do
it, but now it shrinks, and a test that demonstrated this was fixed to
show it shrinks now.