More tests, so far isomorphism checks work
Found that an internal error occurred due to the Vertices
remove_vertex would pop_back the names but not the vertices vector.
In most cases the names size is used as the size of the number of
vertices too, but if iterating the vertex vector directly, we could
see "garbage" at the end which confused the isomorphism
range-detecting logic.

Added several tests, most are recognizing positive isomprophsims, but
I need more tests to show it detects when they are not.  Could use
more methodical test cases.  But a nice start.